% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_to_json.R
\name{quiz_to_json}
\alias{quiz_to_json}
\title{Convert FDS quizzes to json structure.}
\usage{
quiz_to_json(quiz_string)
}
\arguments{
\item{quiz_string}{A string containing one quiz question (be careful to defining it as a string with either `'` or `"` depending on which one is used inside the quiz itself.}
}
\value{
A string containing the quiz in json structure
}
\description{
Convert FDS quizzes to json structure.
}
\examples{
}
